backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.3.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.128.0
fonttools==4.47.0
frozenlist==1.7.0
fsspec==2024.12.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.11
jiter==0.10.0
joblib==1.4.2
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-experimental==0.3.4
langchain-neo4j==0.4.0
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
marshmallow==3.26.1
matplotlib==3.8.2
multidict==6.4.4
mypy_extensions==1.1.0
neo4j==5.28.1
neo4j-graphrag==1.7.0
numpy==1.26.2
ollama==0.5.1
openai==1.87.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
propcache==0.3.2
protobuf==6.33.4
pydantic==2.12.5
pydantic-settings==2.9.1
pydantic_core==2.41.5
pyparsing==3.1.1
pypdf==5.6.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-utils==3.9.1
pytz==2023.3.post1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
urllib3==2.4.0
uvicorn==0.40.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
